Initializaiton setup
Created an initialization scene to have a landing in order to load the correct first scene, in this case main menu. This initialization scene also utilizes the Zenject project context to house persistant data such as the HighScore and LastScore. This also houses the SceneController object which will be the controller that is utilized to load between scenes.